minor top nav refactor
