+++ js_modules/dagster-ui/packages/ui-core/src/app/AppTopNav/AppTopNavRightOfLogo.oss.tsx
@@ -1,0 +1,9 @@
+import {memo} from 'react';
+import {useHistory} from 'react-router-dom';
+
+import {AppTopNavLinks, navLinks} from './AppTopNavLinks';
+
+export const AppTopNavRightOfLogo = memo(() => {
+  const history = useHistory();
+  return <AppTopNavLinks links={navLinks(history)} />;
+});
