+++ js_modules/dagster-ui/packages/ui-core/src/app/AppTopNav/AppTopNavRightOfLogo.tsx
@@ -1,0 +1,3 @@
+import {componentStub} from '../CloudComponentContext';
+
+export const AppTopNavRightOfLogo = componentStub('AppTopNavRightOfLogo');
