+++ js_modules/dagster-ui/packages/ui-core/src/app/__tests__/AppTopNav.test.tsx
@@ -2,7 +2,7 @@ import {MockedProvider} from '@apollo/client/testing';
 import {render, screen} from '@testing-library/react';
 import {MemoryRouter} from 'react-router-dom';
 
-import {AppTopNav} from '../AppTopNav';
+import {AppTopNav} from '../AppTopNav/AppTopNav';
 
 // We don't need to render the search input here.
 jest.mock('../../search/SearchDialog', () => ({
@@ -14,7 +14,7 @@ describe('AppTopNav', () => {
     render(
       <MockedProvider>
         <MemoryRouter>
-          <AppTopNav searchPlaceholder="Test..." rightOfSearchBar={<div>RightOfSearchBar</div>} />
+          <AppTopNav searchPlaceholder="Test..." />
         </MemoryRouter>
       </MockedProvider>,
     );
@@ -25,6 +25,5 @@ describe('AppTopNav', () => {
     expect(screen.getByText('Runs').closest('a')).toHaveAttribute('href', '/runs');
     expect(screen.getByText('Assets').closest('a')).toHaveAttribute('href', '/assets');
     expect(screen.getByText('Deployment').closest('a')).toHaveAttribute('href', '/locations');
-    expect(screen.getByText('RightOfSearchBar')).toBeVisible();
   });
 });
